+++ lib/simple_form/wrappers/root.rb
@@ -24,7 +24,9 @@ module SimpleForm
 
       def html_classes(input, options)
         css = options[:wrapper_class] ? Array.wrap(options[:wrapper_class]) : @defaults[:class]
-        css += SimpleForm.additional_classes_for(:wrapper) { input.additional_classes }
+        css += SimpleForm.additional_classes_for(:wrapper) do
+          input.additional_classes + [input.input_class]
+        end
         css << (options[:wrapper_error_class] || @defaults[:error_class]) if input.has_errors?
         css << (options[:wrapper_hint_class] || @defaults[:hint_class]) if input.has_hint?
         css.compact